# desktop/Byokey/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "byokey-proxy",
        "description": "Bring Your Own Keys \u2014 AI subscription-to-API proxy gateway",
        "contact": {
            "name": "AprilNEA",
            "email": "github@sku.moe"
        },
        "license": {
            "name": "MIT OR Apache-2.0",
            "identifier": "MIT OR Apache-2.0"
        },
        "version": "0.8.0"
    },
    "paths": {
        "/v0/management/accounts": {
            "get": {
                "tags": [
                    "management"
                ],
                "summary": "Lists all accounts for every provider.",
                "description": "# Errors\n\nReturns `ApiError` if the underlying store query fails.",
                "operationId": "accounts_handler",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AccountsResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v0/management/accounts/{provider}/{account_id}": {
            "delete": {
                "tags": [
                    "management"
                ],
                "summary": "Removes a stored account (and its token) for a provider.",
                "description": "# Errors\n\nReturns 400 if the provider is unknown, or `ApiError` on store failure.",
                "operationId": "remove_account_handler",
                "parameters": [
                    {
                        "name": "provider",
                        "in": "path",
                        "description": "Provider identifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "account_id",
                        "in": "path",
                        "description": "Account identifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    }
                }
            }
        },
        "/v0/management/accounts/{provider}/{account_id}/activate": {
            "post": {
                "tags": [
                    "management"
                ],
                "summary": "Switches the active account for a provider.",
                "description": "# Errors\n\nReturns 400 if the provider is unknown, or `ApiError` on store failure.",
                "operationId": "activate_account_handler",
                "parameters": [
                    {
                        "name": "provider",
                        "in": "path",
                        "description": "Provider identifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "account_id",
                        "in": "path",
                        "description": "Account identifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    }
                }
            }
        },
        "/v0/management/ratelimits": {
            "get": {
                "tags": [
                    "management"
                ],
                "summary": "Returns the latest captured rate limit headers for all providers.",
                "operationId": "ratelimits_handler",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RateLimitsResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v0/management/status": {
            "get": {
                "tags": [
                    "management"
                ],
                "summary": "Returns the current server and provider status.",
                "operationId": "status_handler",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StatusResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v0/management/usage": {
            "get": {
                "tags": [
                    "management"
                ],
                "summary": "Returns current in-memory usage counters.",
                "operationId": "usage_handler",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UsageSnapshot"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v0/management/usage/history": {
            "get": {
                "tags": [
                    "management"
                ],
                "summary": "Returns bucketed usage history from the persistent store.",
                "operationId": "usage_history_handler",
                "parameters": [
                    {
                        "name": "from",
                        "in": "path",
                        "description": "Start of the time range (unix timestamp). Defaults to 24 hours ago.",
                        "required": true,
                        "schema": {
                            "type": [
                                "integer",
                                "null"
                            ],
                            "format": "int64"
                        }
                    },
                    {
                        "name": "to",
                        "in": "path",
                        "description": "End of the time range (unix timestamp). Defaults to now.",
                        "required": true,
                        "schema": {
                            "type": [
                                "integer",
                                "null"
                            ],
                            "format": "int64"
                        }
                    },
                    {
                        "name": "model",
                        "in": "path",
                        "description": "Optional model name filter.",
                        "required": true,
                        "schema": {
                            "type": [
                                "string",
                                "null"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UsageHistoryResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/models": {
            "get": {
                "tags": [
                    "management"
                ],
                "summary": "Handles `GET /v1/models` requests.",
                "description": "Returns an OpenAI-compatible model list from the unified registry.\nFor models available on multiple providers, both unqualified (primary)\nand qualified (`provider/model`) forms are listed.",
                "operationId": "list_models",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ModelsResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "AccountDetail": {
                "type": "object",
                "description": "Details for a single stored account.",
                "required": [
                    "account_id",
                    "is_active",
                    "token_state"
                ],
                "properties": {
                    "account_id": {
                        "type": "string"
                    },
                    "expires_at": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "format": "int64",
                        "minimum": 0
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "label": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "token_state": {
                        "$ref": "#/components/schemas/TokenStateDto"
                    }
                }
            },
            "AccountRateLimit": {
                "type": "object",
                "description": "Rate limit snapshot for a single provider account.",
                "required": [
                    "account_id",
                    "snapshot"
                ],
                "properties": {
                    "account_id": {
                        "type": "string"
                    },
                    "snapshot": {
                        "$ref": "#/components/schemas/RateLimitSnapshot"
                    }
                }
            },
            "AccountsResponse": {
                "type": "object",
                "description": "All accounts grouped by provider.",
                "required": [
                    "providers"
                ],
                "properties": {
                    "providers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProviderAccounts"
                        }
                    }
                }
            },
            "AuthStatus": {
                "type": "string",
                "description": "Authentication state for a provider.",
                "enum": [
                    "valid",
                    "expired",
                    "not_configured"
                ]
            },
            "ModelEntry": {
                "type": "object",
                "description": "A single model entry.",
                "required": [
                    "id",
                    "object",
                    "created",
                    "owned_by"
                ],
                "properties": {
                    "created": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "id": {
                        "type": "string"
                    },
                    "object": {
                        "type": "string"
                    },
                    "owned_by": {
                        "type": "string"
                    }
                }
            },
            "ModelStats": {
                "type": "object",
                "description": "Per-model usage counters.",
                "required": [
                    "requests",
                    "success",
                    "failure",
                    "input_tokens",
                    "output_tokens"
                ],
                "properties": {
                    "failure": {
                        "type": "integer",
                        "format": "int64",
                        "minimum": 0
                    },
                    "input_tokens": {
                        "type": "integer",
                        "format": "int64",
                        "minimum": 0
                    },
                    "output_tokens": {
                        "type": "integer",
                        "format": "int64",
                        "minimum": 0
                    },
                    "requests": {
                        "type": "integer",
                        "format": "int64",
                        "minimum": 0
                    },
                    "success": {
                        "type": "integer",
                        "format": "int64",
                        "minimum": 0
                    }
                }
            },
            "ModelsResponse": {
                "type": "object",
                "description": "OpenAI-compatible model list response.",
                "required": [
                    "object",
                    "data"
                ],
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ModelEntry"
                        }
                    },
                    "object": {
                        "type": "string"
                    }
                }
            },
            "ProviderAccounts": {
                "type": "object",
                "description": "Accounts for a single provider.",
                "required": [
                    "id",
                    "display_name",
                    "accounts"
                ],
                "properties": {
                    "accounts": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AccountDetail"
                        }
                    },
                    "display_name": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    }
                }
            },
            "ProviderRateLimits": {
                "type": "object",
                "description": "Rate limit data for a single provider.",
                "required": [
                    "id",
                    "display_name",
                    "accounts"
                ],
                "properties": {
                    "accounts": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AccountRateLimit"
                        }
                    },
                    "display_name": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    }
                }
            },
            "ProviderStatus": {
                "type": "object",
                "description": "Per-provider status summary.",
                "required": [
                    "id",
                    "display_name",
                    "enabled",
                    "auth_status",
                    "models_count"
                ],
                "properties": {
                    "auth_status": {
                        "$ref": "#/components/schemas/AuthStatus"
                    },
                    "display_name": {
                        "type": "string"
                    },
                    "enabled": {
                        "type": "boolean"
                    },
                    "id": {
                        "type": "string"
                    },
                    "models_count": {
                        "type": "integer",
                        "minimum": 0
                    }
                }
            },
            "RateLimitSnapshot": {
                "type": "object",
                "description": "A snapshot of rate limit headers captured from a single upstream response.",
                "required": [
                    "headers",
                    "captured_at"
                ],
                "properties": {
                    "captured_at": {
                        "type": "integer",
                        "format": "int64",
                        "description": "Unix timestamp (seconds) when this snapshot was captured.",
                        "minimum": 0
                    },
                    "headers": {
                        "type": "object",
                        "description": "Raw rate-limit-related header key/value pairs.",
                        "additionalProperties": {
                            "type": "string"
                        },
                        "propertyNames": {
                            "type": "string"
                        }
                    }
                }
            },
            "RateLimitsResponse": {
                "type": "object",
                "description": "Top-level rate limits response.",
                "required": [
                    "providers"
                ],
                "properties": {
                    "providers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProviderRateLimits"
                        }
                    }
                }
            },
            "ServerInfo": {
                "type": "object",
                "description": "Server listen address.",
                "required": [
                    "host",
                    "port"
                ],
                "properties": {
                    "host": {
                        "type": "string"
                    },
                    "port": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 0
                    }
                }
            },
            "StatusResponse": {
                "type": "object",
                "description": "Top-level status response.",
                "required": [
                    "server",
                    "providers"
                ],
                "properties": {
                    "providers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProviderStatus"
                        }
                    },
                    "server": {
                        "$ref": "#/components/schemas/ServerInfo"
                    }
                }
            },
            "TokenStateDto": {
                "type": "string",
                "description": "Token validity state.",
                "enum": [
                    "valid",
                    "expired",
                    "invalid"
                ]
            },
            "UsageBucket": {
                "type": "object",
                "description": "Time-bucketed usage aggregation result.",
                "required": [
                    "period_start",
                    "model",
                    "request_count",
                    "input_tokens",
                    "output_tokens"
                ],
                "properties": {
                    "input_tokens": {
                        "type": "integer",
                        "format": "int64",
                        "minimum": 0
                    },
                    "model": {
                        "type": "string"
                    },
                    "output_tokens": {
                        "type": "integer",
                        "format": "int64",
                        "minimum": 0
                    },
                    "period_start": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "request_count": {
                        "type": "integer",
                        "format": "int64",
                        "minimum": 0
                    }
                }
            },
            "UsageHistoryQuery": {
                "type": "object",
                "description": "Query parameters for the usage history endpoint.",
                "properties": {
                    "from": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "format": "int64",
                        "description": "Start of the time range (unix timestamp). Defaults to 24 hours ago."
                    },
                    "model": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "Optional model name filter."
                    },
                    "to": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "format": "int64",
                        "description": "End of the time range (unix timestamp). Defaults to now."
                    }
                }
            },
            "UsageHistoryResponse": {
                "type": "object",
                "description": "Response for the usage history endpoint.",
                "required": [
                    "from",
                    "to",
                    "bucket_seconds",
                    "buckets"
                ],
                "properties": {
                    "bucket_seconds": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "buckets": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UsageBucket"
                        }
                    },
                    "from": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "to": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "UsageSnapshot": {
                "type": "object",
                "description": "JSON-serializable snapshot of current usage.",
                "required": [
                    "total_requests",
                    "success_requests",
                    "failure_requests",
                    "input_tokens",
                    "output_tokens",
                    "models"
                ],
                "properties": {
                    "failure_requests": {
                        "type": "integer",
                        "format": "int64",
                        "minimum": 0
                    },
                    "input_tokens": {
                        "type": "integer",
                        "format": "int64",
                        "minimum": 0
                    },
                    "models": {
                        "type": "object",
                        "additionalProperties": {
                            "$ref": "#/components/schemas/ModelStats"
                        },
                        "propertyNames": {
                            "type": "string"
                        }
                    },
                    "output_tokens": {
                        "type": "integer",
                        "format": "int64",
                        "minimum": 0
                    },
                    "success_requests": {
                        "type": "integer",
                        "format": "int64",
                        "minimum": 0
                    },
                    "total_requests": {
                        "type": "integer",
                        "format": "int64",
                        "minimum": 0
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "management",
            "description": "Daemon management API"
        }
    ]
}